Removed some warnings;

+++ Engine/Engine/src/graphics/Model.cpp
@@ -87,7 +87,7 @@ void Model::DrawIndexedInstanced(ICamera * cam) const
 	uint32_t renderInstances = 0;
 	for (size_t i = 0; i < m_objectWorld.size(); ++i)
 	{
-		if (ShouldRenderInstance(cam, i))
+		if (ShouldRenderInstance(cam, (uint32_t)i))
 		{
 			data[renderInstances++] = m_objectWorld[i];
 		}
